Prebuilder System
This system will allow SonicTimm to prebuild sections of the levels called roads ahead of time, set if they are rare road sections, how rare rare road sections are, as well as then randomly select and build the portions as the player progresses through the level. There were a few minor changes to other scripts and such to allow for better handling of their systems.