Two player chaos mode
That's what I'm calling this. Because at the moment, it's overly simple. The game is two screens as Timm wanted,  but there is only one player displayed on both screens. Both screens can be bumped to cause the player to jump, but with two players trying to control the jumping at once it would be chaos.

NOT A REAL OR GOOD GAME MODE.
Dev game mode really.