2PlayerMode
Nearly Finished. Updated the camera flip script to work correctly ont he correct camera each time now, and the game at start detects if it is one or two player mode. If it is one player mode it cleans up player two's stuff and modifies the vieport for player one. If two player is set it modifies the vieport for player 2 and 1.