Finishing this quickly.
I've modified the main menu slightly so that it works, changed the game mode to single player and am now focusing on finishing the single player experience to a degree that it can be turned in for now. Included Music manager via a tutorial, though it's overly complicated. It was mostly for practice purposes. Tutorial is cited in code comments.